Add compile hook for doing coverage from ASDF
The compile hook, along with shadowing of in-package, lets
one turn coverage on and off without changing a file.
(needs testing)

Move the sublis form that instruments code into a separate
function.  This will be expanded into something smarter that
dispatches on the car of cons forms, to enable COVER to work
with macros that depend on certain Lisp symbols (like if, when,
etc.) being in their bodies.  An example of that is ITERATE.

Fixed many tests that had inlined defuns.

Added a variable to control if the coverage report is printed
readably or not.  Readable printing tends to pollute it with
package names, which are useless and annoying there.