Extensions to cover to work with a modified version of iterate
The ordinary handling of macros in iterate's code walker will
not work, since it expands them, but iterate does not tolerate
macrolets.   Therefore, special handlers must be invoked for the
c-... macros introduced by cover.

To use the extensions, include the :cover-iterate system.

The extensions to iterate are on the cover branch at my (pfdietz)
repo on github.